+++ core/src/main/kotlin/io/github/kerelape/reikai/core/StructGuard.kt
@@ -1,4 +1,4 @@
-package io.github.kerelape.reikai.guards
+package io.github.kerelape.reikai.core
 
 import io.github.kerelape.reikai.core.Entity
 import io.github.kerelape.reikai.core.EntityWrap
