+++ core/src/main/kotlin/io/github/kerelape/reikai/math/IntegerGuard.kt
@@ -1,8 +1,8 @@
-package io.github.kerelape.reikai.guards
+package io.github.kerelape.reikai.math
 
 import io.github.kerelape.reikai.core.Entity
 import io.github.kerelape.reikai.core.EntityWrap
-import io.github.kerelape.reikai.math.asEntity
+import io.github.kerelape.reikai.core.StructGuard
 import io.github.kerelape.reikai.text.Text
 
 /**
